Load model in bf16, make API base configurable, allow Vercel CORS origins
Fixes OOM/thrashing when running Qwen2.5-0.5B on memory-constrained EC2
instances by loading in its native bf16 precision instead of upcasting
to fp32. Frontend now reads VITE_API_BASE instead of hardcoding
localhost, and the backend accepts requests from any *.vercel.app origin
so the frontend can be deployed separately from the backend.

+++ backend/app/llm.py
@@ -12,7 +12,7 @@ _model = None
 def get_model():
     global _model
     if _model is None:
-        _model = AutoModelForCausalLM.from_pretrained(MODEL_ID, dtype=torch.float32)
+        _model = AutoModelForCausalLM.from_pretrained(MODEL_ID, dtype=torch.bfloat16)
         _model.eval()
     return _model
